requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
openai==1.3.5
pandas==2.1.3
requests==2.31.0
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0

# Development dependencies
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
black==23.11.0
flake8==6.1.0
isort==5.12.0